Refactor into a lib
